Neatened the build_ui python scripts. They use both the command line pyuic command giving full paths to the inputs and outputs and with relative paths to the inputs and outputs as well as using the uic function that can work recursively throuhg a directory but for which there is no set output directory so is not suitable for the current project structure.

+++ build_ui.py
@@ -22,6 +22,7 @@ specific language governing permissions and limitations under the License.
 import os
 import argparse
 import pathlib
+from pathlib import Path
 
 def get_args():
     """
@@ -32,8 +33,9 @@ def get_args():
     """
     parser = argparse.ArgumentParser(
         prog='build_ui',
-        epilog="""Script for automating the compilation of Qt 
+        epilog="""Script for automating the compilation of Qt
         user interfaces created with QDesigner""")
+
 
     parser.add_argument(
         "-d",
@@ -48,6 +50,7 @@ def get_args():
 
     return parser.parse_args()
 
+
 def compile_user_interface(ui_file, py_file):
     """
     convert a qdesigner ui file to a python file
@@ -59,8 +62,16 @@ def compile_user_interface(ui_file, py_file):
         Returns:
             None
     """
+
+
+    # to make the pyuic5 command create relative path for the import functions it creates in the
+    # compiled *_ui.py files it needs to be executed in the form source directory and in the form
+    # pyuic5 some_file_view.ui -o ..\views\some_file_view_ui.py
+    # The source and target can be strings as the command is executed as a string.
     command = "pyuic5 {} -o {}"
+    print(command)
     os.system(command.format(ui_file, py_file))
+
 
 def list_ui_files(directory):
     """
@@ -75,9 +86,10 @@ def list_ui_files(directory):
     files = os.listdir(directory)
     return [f for f in files if f.endswith(".ui")]
 
+
 def list_compiled_ui_files(directory):
     """
-    list all compiled ui files (Ui_*) in a directory
+    list all compiled ui files (*_ui.py) in a directory
 
         Args:
             directory (string) the directory
@@ -86,7 +98,7 @@ def list_compiled_ui_files(directory):
             list of files in directory starting with Ui_
     """
     files = os.listdir(directory)
-    return [f for f in files if f.startswith("Ui_")]
+    return [f for f in files if f.endswith("_ui.py")]
 
 def main():
     """
@@ -100,24 +112,56 @@ def main():
         return
     print("Path: {}".format(path))
 
+    path_in = path.joinpath("resources")
+    if not path_in.is_dir():
+        print("Error {} is not a directory".format(path))
+        return
+    print("Resources Path: {}".format(path_in))
+
+
+    path_out = path.joinpath("views")
+    if not path_out.is_dir():
+        print("Error {} is not a directory".format(path))
+        return
+    print("Views Path: {}".format(path_out))
+
+    cwd_start = os.getcwd()
+    print("cwd_start: ", cwd_start)
+    os.chdir("cgt/resources")
+    cwd_resources = os.getcwd()
+    print("cwd_resources: ", cwd_resources)
+    os.chdir(cwd_start)
+    os.chdir("cgt/views")
+    cwd_views = os.getcwd()
+    print("cwd_views: ", cwd_views)
+
     if args.clean:
-        files = list_compiled_ui_files(path)
-        print("{} files to be removed".format(len(files)))
-        for file in files:
-            target = path.joinpath(file)
+        files_to_remove = list_compiled_ui_files(cwd_views)
+        print("{} files to be removed".format(len(files_to_remove)))
+        #files_to_process = list_compiled_ui_files(cwd_resources)
+        #print("{} files to be processed".format(len(files_to_process)))
+        for file in files_to_remove:
+            target = Path(cwd_views).joinpath(file)
             print("\t{}".format(target))
             target.unlink()
 
-    files = list_ui_files(path)
-    print("{} files to be compiled:".format(len(files)))
-    for file in files:
-        source = path.joinpath(file)
+    files_to_process = list_ui_files(cwd_resources)
+    print("{} files to be compiled:".format(len(files_to_process)))
+    view_relative_path = "../views"
+    for file in files_to_process:
+        print("type(file) ", type(file))
+        source = Path(cwd_resources).joinpath(file)
+        print("source: ", source)
         file = file.rsplit('.', 1)[0]
-        file += ".py"
-        file = "Ui_" + file
-        target = path.joinpath(file)
+        file += "_ui.py"
+# #        file = "Ui_" + file
+        print("File: ", file)
+        #target = path_out.joinpath(file)
+        target = view_relative_path+'\\'+file
+        print("target: ", target)
         print("\t{} => {}".format(source, target))
-        compile_user_interface(source, target)
+        compile_user_interface(file, target)
+        #compile_user_interface(source, target)
 
 if __name__ == "__main__":
     main()
